refact: Split controllers into more subclasses
