+++ src/Panel/Controller/View/PageFileViewController.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace Kirby\Panel\Controller\View;
+
+/**
+ * Controls the view for a page file
+ *
+ * @package   Kirby Panel
+ * @author    Nico Hoffmann <nico@getkirby.com>
+ * @link      https://getkirby.com
+ * @copyright Bastian Allgeier
+ * @license   https://getkirby.com/license
+ * @since     6.0.0
+ */
+class PageFileViewController extends FileViewController
+{
+	public function breadcrumb(): array
+	{
+		$breadcrumb = $this->model->parents()->flip()->values(
+			fn ($parent) => [
+				'label' => $parent->title()->toString(),
+				'link'  => $parent->panel()->url(true),
+			]
+		);
+
+		return [
+			...$breadcrumb,
+			...parent::breadcrumb(),
+		];
+	}
+}
