+++ src/Panel/Controller/View/SiteFileViewController.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Kirby\Panel\Controller\View;
+
+/**
+ * Controls the view for a site file
+ *
+ * @package   Kirby Panel
+ * @author    Nico Hoffmann <nico@getkirby.com>
+ * @link      https://getkirby.com
+ * @copyright Bastian Allgeier
+ * @license   https://getkirby.com/license
+ * @since     6.0.0
+ */
+class SiteFileViewController extends FileViewController
+{
+}
